Move custom wood type registration to TotemicNeoMod
